pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gnc-enrich"
version = "0.1.0"
description = "GnuCash transaction enrichment and categorization tool"
requires-python = ">=3.11"
dependencies = [
    "lxml",
    "Pillow",
    "pillow-heif",
    "pytesseract",
    "scikit-learn",
    "joblib",
    "flask",
    "requests",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
